Created App and fix wait function in workerpool

+++ domain/model/job.go
@@ -5,7 +5,7 @@ import "net/url"
 // CrawlJob represents an url to work on.
 type CrawlJob struct {
 	SeedURL *url.URL
-	Depth   int
+	Depth   uint64
 }
 
 type CrawlResult struct {
